requirements.txt
# TensorFlow is much easier to install using Anaconda, especially
# on Windows or when using a GPU. Please see the installation
# instructions in INSTALL.md


##### Core scientific packages
jupyterlab~=4.0.8
matplotlib>=3.8.2
numpy>=1.25.2
pandas>=2.1.4
scipy~=1.11.3

##### Machine Learning packages
scikit-learn~=1.3.2

# Optional: the XGBoost library is only used in chapter 7
xgboost>=1.7.6

# Optional: the transformers library is only used in chapter 16
transformers~=4.35.2

##### TensorFlow-related packages

# If you have a TF-compatible GPU and you want to enable GPU support, then
# replace tensorflow-serving-api with tensorflow-serving-api-gpu.
# Your GPU must have CUDA Compute Capability 3.5 or higher support, and
# you must install CUDA, cuDNN and more: see tensorflow.org for the detailed
# installation instructions.

tensorflow~=2.14.1
keras-core
# Optional: the TF Serving API library is just needed for chapter 18.
tensorflow-serving-api~=2.14.0  # or tensorflow-serving-api-gpu if gpu

tensorboard~=2.14.1
tensorboard-plugin-profile~=2.14.0
tensorflow-datasets~=4.9.3
tensorflow-hub~=0.15.0

# Used in chapter 10 and 19 for hyperparameter tuning
keras-tuner~=1.4.6

##### Reinforcement Learning library (chapter 18)

# There are a few dependencies you need to install first, check out:
# https://github.com/openai/gym#installing-everything
gymnasium[Box2D,atari,accept-rom-license]>=0.26.2

# WARNING: on Windows, installing Box2D this way requires:
# * Swig: http://www.swig.org/download.html
# * Microsoft C++ Build Tools:
#   https://visualstudio.microsoft.com/visual-cpp-build-tools/
# It's much easier to use Anaconda instead.

##### Image manipulation
Pillow~=10.1.0
graphviz~=0.20.1

##### Google Cloud Platform - used only in chapter 19
google-cloud-aiplatform~=1.36.2
google-cloud-storage~=2.13.0

##### Additional utilities

# Efficient jobs (caching, parallelism, persistence)
joblib~=1.3.2

# Easy http requests
requests~=2.31.0

# Nice utility to diff Jupyter Notebooks.
nbdime>=3.2.1

# May be useful with Pandas for complex "where" clauses (e.g., Pandas
# tutorial).
numexpr~=2.8.7

# Optional: these libraries can be useful in chapter 3, exercise 4.
nltk~=3.8.1
urlextract~=1.8.0

# Optional: tqdm displays nice progress bars, ipywidgets for tqdm's notebook
#           support
tqdm~=4.66.1
ipywidgets~=8.1.1

# Optional: pydot is only used in chapter 10 for tf.keras.utils.plot_model()
pydot~=1.4.2

# Optional: statsmodels is only used in chapter 15 for time series analysis
statsmodels~=0.14.0

# deps implied by the envionment.yml file:
beartype==0.16.4
box2d-py~=2.3
cartopy==0.22.0
duckdb==0.9.2
geopandas==0.14.1
geopy==2.4.1
hypothesis==6.92.1
isort==5.13.2
kaggle==1.5.16
matplotlib_stubs==0.2.0
mypy==1.7.1
nbstripout==0.6.1
pandas-stubs==2.0.3.230814
pre-commit==3.6.0
pyarrow==14.0.2
pytest-cov==4.1.0
ruff==0.1.8
sqlalchemy_stubs==0.4
streamlit==1.29.0
sweetviz==2.3.1
typer==0.9.0
types-cachetools==5.3.0.7
types-colorama==0.4.15.12
types-decorator==5.1.8.4
types-jsonschema==4.20.0.0
types-openpyxl==3.1.0.32
types-pillow==10.1.0.2
types-protobuf==4.24.0.4
types-pygments==2.17.0.0
types-seaborn==0.13.0.8
types-six==1.16.21.9
types-toml==0.10.8.7
uszipcode==1.0.1
watchdog==3.0.0
widgetsnbextension==4.0.9
ydata-profiling==4.6.3
